Change the way we handle morpheme confidence.
This handles the situation where the function is not in the stats database
and we're doing cover-based swapping. It also fixes a think-o with assigning
to the value from an optional rather than the optional itself. Finally, it
properly handles the case where a morpheme does not appear in a given position
when determining confidence.

IP:STAMP